✨ 共通部分がchildrenを取るよう変更
