+++ src/app/_components/parts/CommonDetailsInfo/index.ts
@@ -1,0 +1,1 @@
+export { CommonDetailsInfo } from './CommonDetailsInfo';
