--- src/app/components/parts/DetailsHead/index.ts
@@ -1,1 +1,0 @@
-export { DetailsHead } from './DetailsHead';
